refactor(appTest) 메인 로직 테스트 CustomAppTest / CustomInputTest 두 파일로 분리
Add:
CustomInputTest.js : 입출력 관련 테스트 코드

Changed:
CustomAppTest.js : 입출력 관련 테스트 제거 + 숫자 야구 관련 테스트 저장
